third_party/blink/web_tests/platform/generic/http/tests/loading/preload-append-scan-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame

This test needs to be run in DRT. Preload scanner should see the image resource.
<img src="resources/preload-test.jpg">

PASS
